Refactor front and back office routing separation
